cli/mcq-generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcq-generator"
version = "0.1.0"
description = "Schema-driven MCQ generator powered by DSPy and Mistral AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.1.0",
    "dspy-cli>=0.1.13",
    "pydantic>=2.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]
